move wallet connect consts
