+++ packages/consts/src/walletConnect.ts
